Use a inline worker instead of a file
