--- lib/common.js
@@ -1,3 +1,0 @@
-// We need to make sure that we connect to 'our' worker
-// TODO: this is gross, think of another way
-export const key = '@@symbol:__wasm_api__'
