+++ lib/index.js
@@ -1,17 +1,14 @@
-import { key } from './common'
+import getWorker from './worker-scope'
 
-export default function wasmWorkerAPI (worker, functions) {
-  if (!(worker instanceof Worker))
-    throw new TypeError('worker should be an instance of the Worker')
-
+export default function wasmWorkerAPI (functions, url) {
   if (!Array.isArray(functions))
     throw new TypeError('functions must be an array')
 
   const api = Object.create(null)
-  const secret = Math.random().toString(36).slice(2)
   const types = ['string', 'number', '8', 'U8', '16', 'U16', '32', 'U32', 'F32', 'F64']
+  const worker = getWorker(url)
 
-  let pingPongMsgCounter = 0
+  let id = 0
 
   for (const { name, args: argTypes, ret } of functions) {
     if (!name)
@@ -30,20 +27,12 @@ export default function wasmWorkerAPI (worker, functions) {
       : (...args) => call(name, args)
   }
 
-  const ready = handleMessage(getCurrentId(), (resolve, data) => {
-    if (data[key] !== secret) return
-
-    resolve(api)
-  })
+  const ready = handleMessage(id, (resolve, _) => resolve(api))
 
   worker.postMessage({
-    [key]: secret,
+    url: location.toString(),
     functions
   })
-
-  function getCurrentId () {
-    return `${secret}-${pingPongMsgCounter}`
-  }
 
   function validateReturnType (name, ret) {
     const validReturnType = types.some(type => type === ret)
@@ -82,10 +71,10 @@ export default function wasmWorkerAPI (worker, functions) {
     }
   }
 
-  function handleMessage (id, callback) {
+  function handleMessage (msgId, callback) {
     return new Promise((resolve, reject) => {
       function onMessage ({ data }) {
-        if (data.id !== id) return
+        if (data.id !== msgId) return
 
         worker.removeEventListener('message', onMessage)
 
@@ -119,12 +108,9 @@ export default function wasmWorkerAPI (worker, functions) {
   }
 
   function call (name, args, arrayLength) {
-    pingPongMsgCounter++
-
-    const id = getCurrentId()
-    const promise = handleMessage(id, (resolve, { ret }) => resolve(ret))
-
+    const promise = handleMessage(++id, (resolve, { ret }) => resolve(ret))
     const transferables = []
+
     args && args.length && args.forEach(arg => {
       !!arg.buffer && transferables.push(arg.buffer)
     })
